Fix Blizzard integration bugs surfaced in manual testing
- BlizzardProfileClient: drop dynamic-{region} override on guild
  endpoints — guild API requires profile-{region}, override caused 403s
  that left guild rows empty on the UI.
- CharacterSpecializationMapper: walk the actual response nesting
  (specializations[].loadouts[]) and match the active loadout by
  active_specialization.id; the old flat iteration left talents empty.
  pvp_talent_slots lives on the spec entry, not the loadout.
- BlizzardGameDataClient::getCurrentMythicPlusSeason: move the (int)
  cast outside Cache::remember so stale string values cached from a
  prior run do not break the return type.
- bootstrap/app.php: replace nonexistent everyTwelveHours() with
  twiceDaily() — scheduler was crashing every minute.
- CLAUDE.md: document the per-client namespace invariant, the talent
  source shape, and why docker compose restart horizon is required
  (PHP_OPCACHE_VALIDATE_TIMESTAMPS=0).

+++ app/Blizzard/Client/BlizzardGameDataClient.php
@@ -26,7 +26,7 @@ class BlizzardGameDataClient extends BlizzardClient
             return (int) $override;
         }
 
-        return Cache::remember('blizzard:mythic-plus:current-season', 86400, function () {
+        return (int) Cache::remember('blizzard:mythic-plus:current-season', 86400, function () {
             $response = $this->request()
                 ->get('/data/wow/mythic-keystone/season/index');
